Cleaned up Silkscreen and added caps
I cleaned up the silkscreen for all of the components.
Capacitors were added to the motor driver VCC pins.
More GND Vias were added to reduce noise.
The mosfets were shifted away from the connectors slightly.
The stepper motor connectors were shifted slightly closer to the edge of
the board.
All of the HV pins are labelled as 24v. This board should only be used
with 24v.